Move hearingDetails.ts to root of lib directory
